Split julia support into legacy and current
