+++ repo2docker/buildpacks/__init__.py
@@ -2,6 +2,7 @@ from .base import BuildPack, BaseImage
 from .python import PythonBuildPack
 from .conda import CondaBuildPack
 from .julia import JuliaBuildPack
+from .julialegacy import JuliaLegacyBuildPack
 from .docker import DockerBuildPack
 from .legacy import LegacyBinderDockerBuildPack
 from .r import RBuildPack
